Python: Extract hidden files/dirs by default
Changes the default behaviour of the Python extractor so that hidden
files (and files inside hidden directories) are extracted by default.

Also adds an extractor option, `skip_hidden_files`, which can be set to
`true` in order to revert to the old behaviour.

Finally, I made the logic surrounding what is logged in various cases a
bit more obvious.

Technically this changes the behaviour of the extractor (in that hidden
excluded files will now be logged as `(excluded)`, but I think this
makes more sense anyway.